Initial version of a "crash" uploading system
This will hook into the dispatcher and timer dispatching code and upload a crash dump to a server. There is some conceptual duplication with shoreline, in my cases having a full crash dump is preferred and not a privacy issue (as it is meant to monitor your own deployments). It doesn't change the UIManager default behavior for other/unrelated exceptions.